8178489: Make align functions more type safe and consistent

+++ hotspot/src/share/vm/gc/shared/collectedHeap.cpp
@@ -445,7 +445,7 @@ oop CollectedHeap::new_store_pre_barrier(JavaThread* thread, oop new_obj) {
 }
 
 size_t CollectedHeap::filler_array_hdr_size() {
-  return size_t(align_object_offset(arrayOopDesc::header_size(T_INT))); // align to Long
+  return align_object_offset(arrayOopDesc::header_size(T_INT)); // align to Long
 }
 
 size_t CollectedHeap::filler_array_min_size() {
