+++ hotspot/src/share/vm/gc/shared/space.cpp
@@ -689,7 +689,7 @@ void ContiguousSpace::allocate_temporary_filler(int factor) {
   size = align_object_size(size);
 
   const size_t array_header_size = typeArrayOopDesc::header_size(T_INT);
-  if (size >= (size_t)align_object_size(array_header_size)) {
+  if (size >= align_object_size(array_header_size)) {
     size_t length = (size - array_header_size) * (HeapWordSize / sizeof(jint));
     // allocate uninitialized int array
     typeArrayOop t = (typeArrayOop) allocate(size);
